+++ test/cloud/helloworld/src/main/java/pers/acp/spring/cloud/server/helloworld/hystrix/HelloServerHystrix.java
@@ -1,7 +1,7 @@
 package pers.acp.spring.cloud.server.helloworld.hystrix;
 
-import feign.hystrix.FallbackFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cloud.openfeign.FallbackFactory;
 import org.springframework.stereotype.Component;
 import pers.acp.spring.boot.interfaces.LogAdapter;
 import pers.acp.spring.cloud.server.helloworld.feign.HelloServer;