+++ doc/version_history.md
@@ -13,6 +13,7 @@
 > - acp-spring-cloud-starter
 >   - [Upgrade] AcpCloudResourceServerAutoConfiguration 中关闭 formLogin 登录界面
 >   - [Upgrade] 增加依赖 org.springframework.cloud:spring-cloud-starter-bootstrap
+>   - [Upgrade] 去除ribbon依赖
 > - test
 >   - [Upgrade] logback 配置属性重命名
 >     - logging.pattern.rolling-file-name → logging.logback.rollingpolicy.file-name-pattern
@@ -20,6 +21,7 @@
 >     - logging.file.max-size → logging.logback.rollingpolicy.max-file-size
 >     - logging.file.total-size-cap → logging.logback.rollingpolicy.total-size-cap
 >     - logging.file.max-history → logging.logback.rollingpolicy.max-history
+>   [Upgrade] admin-server、gateway-server 去除 ribbon 依赖
 > - [Upgrade] 升级依赖
 >   - kotlin coroutines 1.4.2
 >   - kotlin 1.4.21