+++ doc/version_history.md
@@ -14,6 +14,7 @@
 >   - [Upgrade] AcpCloudResourceServerAutoConfiguration 中关闭 formLogin 登录界面
 >   - [Upgrade] 增加依赖 org.springframework.cloud:spring-cloud-starter-bootstrap
 >   - [Upgrade] 去除ribbon依赖
+>   - [Upgrade] 更换security依赖
 > - test
 >   - [Upgrade] logback 配置属性重命名
 >     - logging.pattern.rolling-file-name → logging.logback.rollingpolicy.file-name-pattern