+++ doc/version_history.md
@@ -15,6 +15,7 @@
 >   - [Upgrade] 增加依赖 org.springframework.cloud:spring-cloud-starter-bootstrap
 >   - [Upgrade] 去除ribbon依赖
 >   - [Upgrade] 更换security依赖
+>   - [Upgrade] 修改feign FallbackFactory，不再使用netflix-hystrix
 > - test
 >   - [Upgrade] logback 配置属性重命名
 >     - logging.pattern.rolling-file-name → logging.logback.rollingpolicy.file-name-pattern